+++ .gitignore
@@ -1,5 +1,10 @@
 venv/
 __pycache__/
+wanddb/
+.idea/
+.vscode/
+tests/
+checkpoint*/
 
 encoder_directory/
 input_data_files/
