--- DOCKERFILE
@@ -1,16 +1,0 @@
-FROM 763104351884.dkr.ecr.us-east-1.amazonaws.com/pytorch-training:2.0.1-gpu-py310-cu118-ubuntu20.04-ec2
-
-ENV PATH="/opt/ml/code:${PATH}"
-
-# To determin our user code directory
-ENV SAGEMAKER_SUBMIT_DIRECTORY /opt/ml/code
-
-COPY requirements.txt /opt/ml/code/requirements.txt
-
-RUN pip install --no-cache-dir -r /opt/ml/code/requirements.txt
-
-# Store your code
-COPY BigramGPT.py /opt/ml/code/BigramGPT.py
-
-# Defines script entrypoint
-ENV SAGEMAKER_PROGRAM BigramGPT.py
